shared/api: Update API types
