+++ shared/api/instance_override.go
@@ -2,25 +2,12 @@ package api
 
 import (
 	"time"
-
-	"github.com/google/uuid"
 )
 
 // InstanceOverride defines a limited set of instance values that can be overridden as part of the migration process.
 //
 // swagger:model
 type InstanceOverride struct {
-	InstanceOverridePut
-
-	// UUID corresponding to the overridden instance
-	// Example: 26fa4eb7-8d4f-4bf8-9a6a-dd95d166dfad
-	UUID uuid.UUID `json:"uuid" yaml:"uuid"`
-}
-
-// InstanceOverridePut defines the configurable fields of InstanceOverride.
-//
-// swagger:model
-type InstanceOverridePut struct {
 	// The last time this instance override was updated
 	// Example: 2024-11-12 16:15:00 +0000 UTC
 	LastUpdate time.Time `json:"last_update" yaml:"last_update"`
